hoist up internal/app code into internal/
can divide it up later
