+++ internal/playlists.go
@@ -1,4 +1,4 @@
-package app
+package internal
 
 import (
 	"math"
