+++ internal/structs.go
@@ -1,4 +1,4 @@
-package app
+package internal
 
 import "github.com/zhammer/stride-songs/pkg/spotify"
